feat(python-stack): Implementation of RGB image fetching from Earth Engine
Using Flask, we expose a REST API with a single route (/rgb) allowing to
fetch images at a specific date in a specific location. By default (but
this can be parametrized) the service will get all satellite images
taken 3 months before and after the date, and merge them (using the
median of overlapping pixels) to avoid images with too much noise (such
as clouds for example). The scale (in meter per pixel) can also be
parametrized, with a default value of 100.